+++ src/MusicStore/project.json
@@ -4,6 +4,7 @@
     ],
     "description": "Music store application on K",
     "version": "1.0.0-*",
+    "compilationOptions": { "warningsAsErrors": true },
     "dependencies": {
         "Kestrel": "1.0.0-*",
         "Microsoft.AspNet.Server.IIS": "1.0.0-*",
@@ -16,7 +17,7 @@
         "Microsoft.AspNet.StaticFiles": "1.0.0-*",
         "EntityFramework.SqlServer": "7.0.0-*",
         /*For Mono*/
-        "EntityFramework.InMemory":  "7.0.0-*",
+        "EntityFramework.InMemory": "7.0.0-*",
         "Microsoft.Framework.ConfigurationModel.Json": "1.0.0-*",
         "Microsoft.Framework.OptionsModel": "1.0.0-*"
     },
