+++ test/E2ETests/Helpers.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace E2ETests
+{
+    public class Helpers
+    {
+        public static bool RunningOnMono
+        {
+            get
+            {
+                return Type.GetType("Mono.Runtime") != null;
+            }
+        }
+    }
+}
