+++ test/E2ETests/project.json
@@ -1,9 +1,10 @@
 {
+    "compilationOptions": { "warningsAsErrors": true },
     "commands": {
         "test": "Xunit.KRunner"
     },
     "dependencies": {
-        "Microsoft.AspNet.Server.IIS": "1.0.0-*", 
+        "Microsoft.AspNet.Server.IIS": "1.0.0-*",
         "Microsoft.Framework.Runtime.Interfaces": "1.0.0-*",
         "Xunit.KRunner": "1.0.0-*"
     },
